Set action output with comment body
GitHub Actions have the ability to set outputs so that steps in the same
GHA job can retrieve data from previous actions:

https://help.github.com/en/articles/development-tools-for-github-actions#set-an-output-parameter-set-output

These outputs will be available via:
```
${{ steps.*.outputs.comment }}
```

This PR also adds the `action.yml` for defining the metadata of an
action, so that outputs may be exported:
https://help.github.com/en/articles/metadata-syntax-for-github-actions#outputs

This is just a convenient way for consumers to be able to interact with
the results of a particular terraform action. For example, someone might
want to take the comment and post it to their Slack, or perhaps they
want to do some grepping over the output to check for certain resources
being created, etc.