Add flag to skip commenting if there are no changes
Adds a variable: `$TF_SKIP_COMMENT_IF_NO_CHANGES`

If this variable is not true or unset, which is the default case, then
we'll always post the terraform output back to the PR, regardless of if
there are changes or if the terraform plan succeeded or failed. We'll
always post.

This change relies on terraform's -detailed-exitcode functionality:
https://www.terraform.io/docs/commands/plan.html#detailed-exitcode

Fixes hashicorp#29