Add variablized example
Maybe this is too much hand holding, but these actions tend to be relatively repetitive wrt inputs that they need and those inputs should pretty much always be variablized.  This PR provides an example of how to variablize things like the working directory and the TF version so they are managed in a single assignment rather than 2-6 assignments.